+++ R/comparisons.R
@@ -68,7 +68,6 @@ comparisons <- function(model,
                         contrast_numeric = 1,
                         ...) {
 
-
     if (!isTRUE(list(...)[["internal_call"]])) {
         # if `newdata` is a call to `datagrid`, `typical`, or `counterfactual`, insert `model`
         scall <- substitute(newdata)
@@ -80,6 +79,7 @@ comparisons <- function(model,
             }
         }
         newdata <- sanity_newdata(model, newdata)
+        vcov <- sanitize_vcov(model, vcov)
     }
 
     # TODO: don't run sanity checks if this is an internal call. But
@@ -88,7 +88,6 @@ comparisons <- function(model,
     model <- sanity_model(model = model, ...)
     sanity_type(model = model, type = type)
     variables <- unlist(sanity_variables(model = model, newdata = newdata, variables = variables)[["conditional"]])
-    vcov <- sanitize_vcov(model, vcov)
     checkmate::assert_choice(contrast_factor, choices = c("reference", "sequential", "pairwise"))
     checkmate::assert(
         checkmate::check_numeric(contrast_numeric, min.len = 1, max.len = 2),