+++ NEWS.md
@@ -1,5 +1,12 @@
 # marginaleffects 0.3.5
 
+New function:
+
+* `comparisons()` computes contrasts
+
+Misc:
+
+* Speed optimizations
 * More robust handling of in-formula functions: factor(), strata(), mo()
 * Do not overwrite user's `ggplot2::theme_set()` call